Make differentiabilty property of some function concepts optional
this patch extends the concepts LocalFunction, GridFunction and GridViewFunction
with an additional template parameter to turn off the check for
differentiability, which is useful if you don't require the function to be
differentiable. More importantly, this allows the polymorphic wrappers for these
functions to wrap non-differentiable functions by relaxing their concept checks.

The corresponding compile-time functions for checking the concepts are modified
accordingly.