Implement caching of local DOFs in bind.
We originally wanted to implement this cache but never did this.
I implemented a benchmarking computing the L^2 norm of FE
function with Lagrange-DG basis and various order for the DG
space and the quadrature rule.

It turned out that caching is neven harmful (even for low order).
But it may lead to minor improvements (up 10%) if the quadrature
order (and thus the number of local evaluations) is large.

In order to work properly for proxy-containers this uses
`AutonomousValue` to determine the storage type in the cache.