Make DiscreteGlobalBasisFunction return values in world coordinates
Previously, DiscreteGlobalBasisFunction objects return function values
as they are handed out by the LocalBasis objects.  This is always
the coordinate system of the reference element, which is not what
is desired.  This does not pose a problem for affine families
of finite elements, because there the two coordinate systems
coincide. However, for non-affine FE like Raviart-Thomas finite
elements, an additional transformation is missing (the Piola transform
in case of the RT-FE).  This patch introduces a callback in the
DiscreteGlobalBasisFunction for the correct transformation, and
implements it for the RaviartThomasBasis implementation.