Make the Raviart-Thomas basis apply the Piola transform
We do this by introducing a new LocalFiniteElement implementation
called GlobalValuedLocalFiniteElement.  This new implementation wraps
existing LocalFiniteElement implementations, but applies a
transformation to the evaluated shape function values (and
derivatives etc.).  The transformation is a template parameter
and can be replaced.  Currently only the contravariant Piola transform
(for H(div) spaces) is implemented.

Note that while this patch updates the poisson-mfem example,
the example nevertheless produces a slightly different result now.
The reason is that the example uses the 'interpolate' method,
which did not return the correct values for Raviart-Thomas elements
before this patch.