[bugfix] Implement proper copy of BSplineLFE
Since the local basis stores a reference to the LFE,
we must implement a copy constructor to avoid a dangling
reerence. Due to guaranteed copy elision there is no
copy when storing the `BSplineNode` in the `LocalView`.
Hence this problem does not show up in practice.

The situation is different when using a power of
a `BSplineBasis`. Then the `PowerNode` stores several
copies. Hence accessing the basis fails with a segfault
due to the dangling reference.

This is the reason why the extended basis check failed
with a segfault for the `BSplineBasis`.