[test] Factor out local continuity check on intersection
The global continuity check now does all the generic stuff:
* Loop over elements
* Find neighbors
* Loop over all inside leaf nodes
* Find corresponding outside leaf node
* Match inside DOFs to outside DOFs using global indices

Using this information it executes a local continuity check.
The latter gets passed the intersection, inside- and outside-node,
and the map from inside to outside DOF (if existent). The result
of the local check should be a random access range indicating
for each (inside) basis function if is is continuous across the
intersection.