[test] Pass tree path to local continuity check
This was an oversight in the initial version of the continuity check:
For a nontrivial tree you may want to use different continuity checks
for different leaf nodes. In part this could be achieved by switching
with the node type. But this lacks some information, because the same
(local) node may have different global continuity properties. This
is e.g. the case for the `LagrangeBasis` and the `LagrangeDGBasis`.
Hence we miss some information if we do not pass the tree path.