Introduce PeriodicBasis
PeriodicBasis is a meta basis---a basis that is parametrized
with another basis.  In a PeriodicBasis, global degrees of freedom
of the host basis can be grouped into equivalence classes,
which are then treated as single global degrees of freedom by the
PeriodicBasis.  This allows, in particular, to implement periodic
boundary conditions for discretizations without intersection
integrals.