Introduce PeriodicBasis
PeriodicBasis is a meta basis---a basis that is parametrized
with another basis.  In a PeriodicBasis, global degrees of freedom
of the host basis can be grouped into equivalence classes,
which are then treated as single global degrees of freedom by the
PeriodicBasis.  This allows, in particular, to implement periodic
boundary conditions for discretizations without intersection
integrals.

As there was some controversy about whether such a basis is really
needed the implementation is in the namespace
Dune::Functions::Experimental for now.  That will the code some
more exposure.

Thanks go to Carsten Gräser, who wrote a lot of code in this commit.