Implement a Nedelec basis (of the first kind)
This patch adds the implementation of a Nedelec basis.  The interface
allows bases of any order, grids in 2d and 3d, and with simplex and cube elements.
However, this initial implementation can only do the first-order basis
on triangle grids.

There are two types of Nedelec bases, known as the Nedelec bases of the
first and second kind.  The implementation in this patch has the 'kind'
as a template parameter, which can have values 1 or 2.  However, only
the case kind==1 is implemented right now.