[RaviartThomas] fix 3rd-order for quads and 2nd-order for hexas
The local finite element selection mechanism in RaviartThomasBasis would
only work properly if a Raviart-Thomas local finite element of given
order and space dimension existed both for cubes and simplices.
This is not the case for 3rd-order RT elements in 2d and 2d-order
RT elements in 3d:  They currently exist for cubes, but not for
simplices.  This patch makes the element selection logic more
flexible, and thereby implements these to new cases.