Add support for PreBasis::indices() to DefaultLocalView
This provides support for `PreBasis` implementations directly
providing an `indices(node,iterator)` method, while still
supporting the old `NodeIndexSet`-based interface.

This also adds a concept check for the new intreface and
a fallback method `Impl::preBasisIndices(preBasis,node,it)`
that forwards to the old interface if the new one is not
implemented. The fallback method can be used inside of
a meta-prebasis as a replacement of `preBasis.indices(node,it)`
for a transition period.

Passing a prebasis not providing the new interface to
`DefaultGlobalBasis` is deprecated.