[bugfix] Fix caching in DiscreteGlobalBasisFunction
This fixes an issue with caching in `DiscreteGlobalBasisFunction`
when using a subspace basis. For a subspace basis the size-of-tree and
index-within-tree do not coincide with size-of-localview
and index-within-localview (aka localindex), repectively.
Since this was not handled carefully which led to memory corruption
and thus valgrind errors and non-deterministic values whenever the
subspace was not starting with local index 0.

Now the cache has the size of the full local view but only
entries associated to the tree (i.e. subspace) are actually
written and read.