Add ComposedGridFunction
For given inner grid functions `g0`, ..., `gn` and an
outer function `f` this implements the grid function
mapping `x`  to `f(g0(x), ..., gn(x))`. It is assumed
that the grid functions are defined on the same
`EntitySet` and that their range type can be passed to
the outer function.

All functions are stored by value. Storing references
is possible using passing `std::ref(f)`. To resolve
extended methods internally this makes use of the
`Dune::Functions::resolveRef()` utility.