[python] Update global basis creation in python
The old solution was to first construct the nested `PreBasisFactory`
object relying on the fact that this does not store any information.
With the modified `PreBasis` interface this is hugely simplified,
since we can directly construct the `PreBasis` type from the `GridView`.
Now we can directly construct the global basis from the `GridView`.

This assumes that the `PreBasis` can be constructed from the `GridView`
which is now also true for the `CompositePreBasis`. Notice that this
does only work if the `PreBasis` does not depend on dynamic data
except the `GridView`. However this assumption was also made implicitly
before but in a much more hidden way. Furthermore it is satisfied for
`LagrangePreBasis` which is the only one supported currently.