Avoid nesting of SubspaceBasis by resolving nested ones
While it was technically allowed and possible to construct a nested
`SubspaceBasis` this is often not intended. This patch avoids such
nesting by joining the inner and outer `prefixPath` and applying it
to the root of the root.

This also adds deduction guids such that subspaceBasis() and
SubspaceBasis() yield the same result.