[bugfix] Avoid implicit conversion to ComposedGridFunction
Before this patch, the constructor of `ComposedGridFunction`
is available for implicit conversion during overload resolution
(but would not compile). Since there is also a function
`localFunction(ComposedGridFunction)` this makes
`localFunction(t)` valid for any `t`. While this will
in general not compile, the overload is still available
and thus makes `t` look like a grid function for concepts
checks e.g. in dune-vtk.

We could make the constructor explicit, but since at least
one inner function is needed, we can also check for this.