move shared data into a separate struct
This allows the LocalFunction to exist even when the associated
DiscreteGlobalBasisFunction has been removed.

It furthermore is expected to ease addition of a derivative: the
derivative's global and local functions have to share the same data
and it is easier to use if there are no shared lifetime requirements
between all of these.