Specialize FlatVectorBackend for std::array
This adds a specialization for `std::array` and some improvements
to hand out size statically. Notice that this only works if
the entries of the `array` have a statically known size.
Ideally one would implement a recursive `FlatVectorBackend`
but I can't get this to compile (with a 'use of ... before decution
of auto' error).