Simplify type when nesting ShiftedDynamicMultiIndex
This adds an overload for `shiftedDynamicMultiIndex()` that
ensures that shifting a multi-index twice will not produce
a nested `ShiftedDynamicMultiIndex` but add the offsets.
While the behaviour of the resulting multi-index is the same,
the type is simplified which eases reading error messages
and debug output, because you get
`ShiftedDynamicMultiIndex<Index,3>` instead of
`ShiftedDynamicMultiIndex<ShiftedDynamicMultiIndex<ShiftedDynamicMultiIndex<Index,1>, 1>, 1>`.