Deprecate class HierarchicVectorWrapper and corresponding header
This was superseeded by `istlVectorBackend(v)`. The difference between
both is, that `HierarchicVectorWrapper` allows to specify the entry
type used to terminate index resolution. In contrast `istlVectorBackend()`
will always terminate for scalar entries. This seems to give the
freedom to use vector-valued coefficients. However, we deliberately
decided to only support scalar coefficients in dune-functions rendering
this feature useless.

(Historically we supported vector-valued coefficients to ease the
transition from the old bases in dune-fufem where this behaviour
was the only way to get vector-valued ansatz spaces.)