[test]Extend basis index consistency test
Now the test first computes all possible prefixes and
corresponding sizes from the indices provided by the
basis. Then consistency of the computed values with
ones returned by `basis.size(prefix)` is checked.

This extends the coverage, because the previous version
of the test only ensured that `basis.size(prefix)` is not
to small. Also it makes the test and the error messages
easier to understand.