[test] Add test of global interpolate() for Hermite bases
This adds the following:

* A general `checkInterpolateConsistency()` utility for testing
  if global `interpolate()` and `DiscreteGlobalBasisFunction`
  are consistent.
* An implementation of a 1d and 2d cubic Hermite basis for testing.
* A new test that applies `checkInterpolateConsistency()`
  to the `CubicHermitePreBasis` to test the updated `interpolate()`
  method.

The `CubicHermitePreBasis` implementation is currently only
used for testing and thus contained in the test directory.
However it is fully functional in 1d and 2d and also brings all
the boiler-plate code needed for general dimension. Hence
we could also add it as real implementation. Internally
it uses a new `LeafPreBasisMapperMixIn` class, that allows
to simplify the `Prebasis` significantly, if index computations
can be done using a `MultipleCodimMultipleGeometryMapper`.