Deprecate namespace BasisBuilder
This was renamed to `BasisFactory` long ago. Unfortunately
we never explicitly deprecated it and it was still used in
our own tests and examples.

While we could not deprecate the namespace when we renamed it,
it's now possible to adds a `[[deprecated]]` attribute and
recent compilers even warn about this.