[python][examples] Add a Python implementation of the mixed Poisson example
This example program does the same as poisson-mfem.cc, but using
the Python interface.  The only difference is the type of the stiffness
matrix.  The C++ example uses a 2x2 matrix of BCRSMatrix objects,
but that is not easy to replicate in Python unless we start writing
Python bindings for dune-istl.  This new Python example uses
a single flat matrix instead.