[python] Fix interpolate by allowing different coefficient and range type
The recent change still does not allow for vector-valued bases,
because the python `interpolate` wrapper hardwires that coefficient
and range type must be the same.
This patch introduces separate template parameters to solve the problem.

Notice that in the vector-valued case, with boolean entries,
using `FieldVector` does not work with numpy bool-arrays,
but `std::array` doesa