[python] Provide missing tree bindings
So far `node.localIndex(i)` was only forwarded to python
for leaf nodes. In fact this method exists for all nodes.
This patch adds the method together with the furthermore
missing `size()` aka `__len__()` and `treeIndex()` methods.

These exist for non-leaf nodes for good reason: They allow
to cheaply deduce the index range for a subtree without
having to traverse is. This is e.g. needed when working
with (or when checking) with `SubspaceBasis` where
local indices are consecutive but not zero bases and where
we have `localView.tree().size() <= localView.size()`
unless the subspace is trivial.