Ensure proper type range type in contravariant Piola-transfomation
We should only assume that the range type of the function passed
in for interpolation is a compatible container. This does not
mean, that we can use it for computations (other than `mv`, `mtv`
which only use component assignment).

Thus we have to deduce a proper range type. As a first solution
we can use the local coordinate type. This is not the most flexible
solution, but I'm not aware of a single use case where this would
currently fail.