[python] Remove non-standard shortcut for vector-valued Lagrange bases
The Python interface had a shortcut that allowed to create vector-valued
Lagrange bases by an optional 'dimRange' parameter of the Lagrange basis
constructor. The shortcut was merely syntactic sugar: Internally a power basis
was used.

Since the C++ interface doesn't have this extra 'dimRange' parameter,
and since no other bases in Python has it either, this commit removes
the 'dimRange' parameter for the sake of consistency.