[python] Initial testing of the individual nodes of a basis tree
This commit mainly adds the recursive loop over the nodes of a tree.
Doing this it tests a few methods: isLeaf & Co, and the degree() method.
More thorough testing will have to come in later commits.

The 'degree' method existed but had to be fixed: In C++ it returns
std::integral_constant<std::size_t>, which is not automatically
converted by pybind11.  (This is deliberate, see

  https://github.com/pybind/pybind11/pull/2328

) Therefore this commit adds an explicit cast to the bindings
of that method.