[examples] Fix sign of pressure in Stokes example
While the exact sign is not important mathematically,
the example used the physically wrong sign. Thus this
is also uncommon in literature. The physically correct form
of the Stokes equation is `-\Delta u + \grad p = f` which,
after partial integration leads to a `-div(v)p` term in the
weak form.

This issues exists since the example was added. Maybe the
initial version of the example was implemented based on the
book by Braess. There, indeed the uncommon sign is used for
the Stokes as well as for the Navier-Stokes equation.

No surprisinly, the uncommon convention leads to confusing
results. E.g. for the lid driven cavity flow, the pressure
decreases towards the corner where the fluid is driven to.
Similarly, for Poiseuille flow in a pipe the fluid is driven
towards the high pressure side.