Relax range type deduction in ComposedGridFunction
Use `localFunction(f)(localX)` instead of `f(globalX)`
for the wrapped grid function `f` when deducing the
range type in `ComposedGridFunction`. This is more
flexible in case the wrapped function does not implement
global evaluation, but only the local one. Since we wrap
a `GridFunction` the latter must always exist.