[bugfix] Don't use mv and mtv with raw random access containers
The Dune matrix types do not support `MatrixType::mv(x,y)` passing
plain random access containers as vectors but require `x.N()` and
`y.N()` if bound checks are enabled.
As a fix we introduce an internal `Impl::DensseVectorView` class
that provides the `DenseVector` interface for generic random
access containers.

Notice that this class is not tested for general use and should
be replaced by `DenseVectorSpan` once this was added to dune-common
and is sufficiently generic for beeing used here.