Delete copy constructor of wrapper
Since `Impl::DenseVectorView` is just a wrapper around
a refence/pointer to a raw container, having a copy
constructor does not make sense. Unless we introduce
dynamic memory management the only possibility is to
either let a copy-constructed `DenseVectorView` to the
same storage or to `nullptr`. Both are not what one would
expect. Assume e.g. the following function expecting a
matrix and vector argument gets a `DenseVectorView` as `x`:

```cpp
template<class A, class X>
auto foo(const A& a, const X& x) {
  auto y = x; // Assume that this copies the internal pointer,
  a.mv(x, y); // then this changes the storage of x when writing to y.
  return x*y;
}
```