Add makeContainer<T>(containerDescriptor) utility
This is similar to `makeISTLVector<T>(containerDescriptor)`
but hands out plain random access containers without arithmetic
operators. The nested containers are composed by `std::vector`,
`std::array`, and `Dune::TupleVector`. In contrast to `makeISTLVector()`
there is no default value for the type `T`. Furthermore the entries
of type `T` are not intinialized by `0` but default initialized.
In order to initialize to a specific value, the latter can be passed
as additional parameter using `makeContainer(containerDescriptor,
value)`.