Add utility GeometryInAncestor
This class represents the geometric embedding
of an element into an acestor by chaining
`geometryInFather()` objects. This information
is computed from a given fine element and a
predicate. The tranversal towards the ancestors
stops at the first element where the predicate
is not satisfied. This element is considered
as coarse element.

While an alternative interface where the coarse
element is the last one satisfyin the predicate
seems to be simpler at first, the one proposed
here is nicer for several reasons:

* If the element range only consists of elements
  satisfying the predicate, we may end up with
  an empty range if the starting element does not
  satisfy it. With the proposed aproach we end up
  with the identity on the fine element.
* There are several applications where the proposed
  interface is very natural, e.g. traversing towards
  the first element that persists after refinement.
* A criterion that first needs to obtain an ancestor
  before it can be decided if it should be included
  is more expensive.