Add utility class CoarseGridViewFunction
Assume that for a given `gridView` the function
`f` is a `GridFunction` defined on some entity
set of the same grid which is coarser than the
`gridView` in the sense that each element in the
`gridView` has an ancestor `e` satisfying
`f.entitySet().contains(e)`.
Then this utility allows to represent `f` as
a `GridViewFunction` on the `gridView` using

```cpp
auto cgvf = CoarseGridViewFunction(f, gridView);
```

When bound to an element from the `gridView`,
this internally searches for an ancestor where
`f` is defined and translates local coordinates.