Add utility class FineGridViewFunction
Assume that for a given `gridView` the function
`f` is a `GridFunction` defined on some entity
set of the same grid which is finer than the
`gridView` in the sense that each element in the
`gridView` is covered by (not necessarily direct)
descendent elements `e` satisfying `f.entitySet().contains(e)`.
Then this utility allows to represent `f` as
a `GridViewFunction` on the `gridView` using

```cpp
auto fgvf = FineGridViewFunction(f, gridView);
```

When bound to an element `e` from the `gridView`
and evaluated locally at some point `x`,
this internally searches for a descendet element
in `f.entitySet()` that contains `x` and
translates local coordinates.