[bugfix][test] Update basis after grid was modified
The test creates the basis, refines the grid and then
uses the basis. This is undefined behaviour and just
happened to work accidentally because the `LangrangeBasis`
implementation computes the dimension on the fly instead
of relying on a mapper that precomputes its size.

The correct behavious is to either create or update the
basis after refinement.