[python][cleanup] Don't specialize DimRange for nodes
The `DimRange` template class from dune-common is used
to spezialize for range types in order to specify the
corresponding range dimension. Specializing `DimRange`
for a node seems to be rather odd. Furthermore this is
only used locally in `registerBasisType()`. Hence we
effectively abuse `DimRange` to sum up the leaf node
range dimensions recursively.

This patch replaces the construction by a simple
recursive `constexpr` function and by the way replaces
the clumsy check for node tags by `is(Leaf|Power|Composite)`
and introduces a static assertion for the non-supported
dynamic power case.