Inform users of pending change of default index merging strategies
Using the methods `power` or `composite` to construct function space
basis trees without an explicit index merging strategy argument
will now issue a warning. This warning informs that the default
index merging strategy will change after the 2.11 release.
More specifically, it will change from `BlockedLexicographic`
to `FlatLexicographic` for the `composite` method, and from
`BlockedInterleaved` to `FlatInterleaved` for the `power` method.
The consequence is that default indexing is always flat,
which is the most common form of indexing.

To ensure a smooth transition we suggest to not rely on default
index merging strategies until 2.11 has been released.