Relax assumptions on bound nodes in utilities
Introduce a check for `node.empty()` in the utilities
acting on global bases. The check is done before the `node` is
used otherwise. This allows to use nodes that have size zero
but lead to UB when other methods than `size()` and `localIndex()`
are accessed which is an important requirement for beeing able
to implement nodes restricted to sub-domains without having
an expensive do-nothing-implementation in any local finite element.

This adjusts the utilities to the relaxed guarantees propose in