[subdomain] Add SubDomainInterface and SubDomainSkeleton class
These represents the interface between two subdomains
and the skeleton of a `SubDomain`
and conceptually represents a ranges of intersections.
They allows to iterate over the intersections and provides
a contains method to check if an intersection is between
the subdomains or in the skeleton, respectively.

Notice that the iterator will visit intersections only
once with a well defined orientation while the `contains()`
method does not care for the orientation.