[bugfix] Implement deep copy of DefaultLocalView
Since there was no explicit copy constructor for
the `DefaultLocalView`, the generated one is used.
Due to the `shared_ptr` storage in the dune-typetree
nodes, this does a (partial) shallow copy of the tree.
To fix this we create a now create a new tree in and
bind it if needed.

Notice that the explicit copy constructor might become
obsolete again, one we got rid of the dune-typetree
node base classes.