[test] Test higher order LagrangeBasis in 2d and 3d
The test now also checks in 2d and 3d with higher order 1,2,3,4,5
to ensure that the basis functions on faces are permuted properly.
While a simplex grid created by the `StructuredGridFactory` already
contains twisted faces in 2d and 3d, this is not the case for
cube grids. Hence we explicitly construct and refine an initial
grid with two twisted cubes. After local refinement the latter
will also contain prism and pyramid elements.

For the latter we only test orders 1 and 2 since the respective
`LocalFiniteElement` implementations do not support higher order.
Once they have been extended, the permutation code of the global
basis should work out of the box.