Move the `squeeze` helper methods into a separate file
Previously, they were contained in the cubichermitebasis.hh file,
but they are used by other bases as well.

While moving, the methods are renamed to `squeezeTensor`, to make
them a little less generic. Moving and renaming the methods is not
an interface change, because they are all contained in the `Impl`
namespace.