Polynomial: Don't add coefficient if it is integral_constant<0>
Adding zeros is likely to be optimized away by the compiler, but not
when using AD systems like ADOL-C: There the +0 will end up on the tape
even with an optimizing compiler.  Therefore, guard for this case by
adding coefficients in the Horner scheme only if they are not
integral_constant<0>.

Yes, this is a small corner case, but it doesn't hurt the general case.
The code gets a tiny bit more complicated, but the complication is
confined to a small part of the Polynomial class.