raviartthomasbasistest: Create grid via a custom grid factory
This is a preparation for also testing bases after the grid view has
been updated.  If the testRaviartThomasBasis is to modify the grid while
testing, then either the calling code has to create a new grid every
time that testRaviartThomasBasis is called, or testRaviartThomasBasis
has to construct the grid object itself. To implement the latter, which
seems cleaner to me, to calling code describes what grid to construct by
handing over a small factory object.