[cleanup] Use TreePath instead of HybridTreePath
The name `TreePath` was reintroduced as alias for `HybridTreePath`
and only the former is intended to be moved to dune-common, while
the latter persists in dune-typetree. Hence we would have to include
`dune/typetree/treepath.hh` instead of `dune/common/treepath.hh`
to use `HybridTreePath`.

Instead, of this, we now consistently use `TreePath` in order
to prepare dropping the dune-typetree dependency.